Move x509 cert logic to sub-directory
The x509 certificate logic has been moved to a dedicated sub-directory
to better group the coming CRL and certificate chain logic.
