+++ verifier/src/lib.rs
@@ -4,19 +4,16 @@
 #![deny(missing_docs, missing_debug_implementations, unsafe_code)]
 #![no_std]
 
-#[cfg(feature = "alloc")]
-mod certs;
 mod report_body;
 mod struct_name;
+#[cfg(feature = "alloc")]
+mod x509;
 
 pub use report_body::{
     AttributesVerifier, ConfigIdVerifier, ConfigSvnVerifier, CpuSvnVerifier,
     ExtendedProductIdVerifier, FamilyIdVerifier, IsvProductIdVerifier, IsvSvnVerifier,
     MiscellaneousSelectVerifier, MrEnclaveVerifier, MrSignerVerifier, ReportDataVerifier,
 };
-
-#[cfg(feature = "alloc")]
-pub use certs::{Error as CertificateError, UnverifiedCertificate, VerifiedCertificate};
 
 use core::fmt::{Debug, Display, Formatter};
 use mc_sgx_core_types::{
